Put segmentize into a source file
